scripts/requirements_deploy.txt
asyncpg
requests
supabase
pydantic-settings
python-dotenv
